Repair missing 7bit name part in rfc2047 on arm and mips architecture.
The problem was, that on arm* and mips* architectures the fetchaddr tests
failed, since from "Günther family1" the " family1" was stripped on these
architectures (amd64, i386, ppc64el and s390x were okay).
This was based on the fact, that strfcpy (and strncpy) on these
architectures didn't copy anything.
Using snprintf instead fixes the issue.